Rename BT Ports scxml declarations
